Move accessory file into subdir
